refactor: split the binary store out of storage.rs (1037 -> 791 lines)
The read guard pushed `storage.rs` past the repository's 1000-line limit
again. `storage_binary.rs` takes `BinaryTokenStore`, its file fingerprint and
its `TokenStore` impl -- the file-mapped doublets projection, which is a
coherent unit and the part this branch has been changing.

No behaviour change; 52 suites pass on the toolchain CI pins.
